+++ changelog.rst
@@ -8,12 +8,12 @@ dev
 - TextPositionSelector no longer has real_start and real_end that can differ from start and end
 - Selectors and TextPositionSets are Pydantic models
 - TextSelector is Pydantic model for either Quote or Position Selector
-- add type field to SelectorSchema
+- split Marshmallow SelectorSchema into Quote and Position schemas
 - update type annotations for TextPositionSelector.from_range
 - add TestQuoteSelector.as_unique_position method
 - TextPositionSet can include TextQuoteSelectors
 - add convert_quotes_to_positions method to TextPositionSet
-- replace `selectors` field with `positions` and `quotes`
+- replace `TextPositionSet.selectors` field with `positions` and `quotes`
 
 0.5.3 (2021-08-11)
 ------------------